move manager file of setting json
