+++ data_gorvernance/library/task_director.py
@@ -6,50 +6,9 @@ from IPython.display import display
 from IPython.core.display import Javascript
 
 from .utils.html.button import create_button
-from .subflow.status import StatusFile, SubflowStatus
 from .utils.config import path_config, message as msg_config
 from .utils.save import TaskSave
-from .main_menu.research_flow_status import ResearchFlowStatusOperater
-
-
-def get_subflow_type_and_id(working_file_path: str):
-    """ノートブックのパスを受け取ってsubflowの種別とidを返す
-
-    Args:
-        working_file_path (str): ノートブックのパス
-
-    Raises:
-        ValueError: working_file_pathにsubflow_typeが含まれない場合
-
-    Returns:
-        str: サブフロー種別
-        str: サブフローID（無い場合はNone）
-    """
-    parts = os.path.normpath(working_file_path).split(os.sep)
-    target_directory = path_config.RESEARCHFLOW
-    subflow_type = ""
-    subflow_id = None
-
-    try:
-        index = parts.index(target_directory)
-    except:
-        raise
-
-    if index < len(parts) - 1:
-        subflow_type = parts[index + 1]
-    else:
-        raise ValueError
-
-    if index + 2 < len(parts) - 1:
-        abs_root = path_config.get_abs_root_form_working_dg_file_path(working_file_path)
-        id_list = ResearchFlowStatusOperater(
-                    path_config.get_research_flow_status_file_path(abs_root)
-                ).get_subflow_ids(subflow_type)
-        dir_name = parts[index + 2]
-        if dir_name in id_list:
-            subflow_id = dir_name
-
-    return subflow_type, subflow_id
+from .utils.setting import get_subflow_type_and_id, SubflowStatusFile, SubflowStatus
 
 
 def get_return_sub_flow_menu_relative_url_path(working_file_path: str)->str:
@@ -99,7 +58,7 @@ class TaskDirector(TaskSave):
             task_name (str): [タスク名]
         """
         # タスク開始によるサブフローステータス管理JSONの更新
-        sf = StatusFile(self._sub_flow_status_file_path)
+        sf = SubflowStatusFile(self._sub_flow_status_file_path)
         sf_status: SubflowStatus = sf.read()
         sf_status.doing_task_by_task_name(task_name, os.environ["JUPYTERHUB_SERVER_NAME"])
         # 更新内容を記録する。
@@ -111,7 +70,7 @@ class TaskDirector(TaskSave):
         Args:
             script_file_name (str): [タスク名]
         """
-        sf = StatusFile(self._sub_flow_status_file_path)
+        sf = SubflowStatusFile(self._sub_flow_status_file_path)
         sf_status: SubflowStatus = sf.read()
         sf_status.completed_task_by_task_name(task_name, os.environ["JUPYTERHUB_SERVER_NAME"])
         sf.write(sf_status)
