+++ data_gorvernance/library/utils/config/param.py
@@ -1,6 +1,7 @@
 import os
-from ...utils.file import JsonFile
 from pathlib import Path
+
+from ...utils.file import JsonFile
 
 # param.jsonのファイルパス
 script_dir_path = os.path.dirname(__file__)
