+++ data_gorvernance/library/utils/nb_file.py
@@ -1,10 +1,11 @@
-import json
+import os
 from typing import Any
-from .file import File
+
 import nbformat
+
+from .file import File
 from .config import message as msg_config
 from .html import security
-import os
 
 class NbFile(File):
 
@@ -21,7 +22,7 @@ class NbFile(File):
         """
         notebook = self.read()
         for cell in notebook.cells:
-             if cell.cell_type == 'markdown' and msg_config.get('DEFAULT', 'param_subflow_name') in cell.source:
+            if cell.cell_type == 'markdown' and msg_config.get('DEFAULT', 'param_subflow_name') in cell.source:
                 cell.source = cell.source.replace(msg_config.get('DEFAULT', 'param_subflow_name'), security.escape_html_text(subflow_name))
         self.write(notebook)
